Creation PropertyController to keep the mainController cleaner
